CSS: Add subcell-size variable and use it to define toolbar height
According to the HIG [1], the value is 15 pixels for the original XO,
and according to sugar-artwork code, the value for 96 DPI is 11 pixels
(72%).  And the latter is the current CSS being developed.  But we are
ready to start supporting other pixel densities, outputting different
CSS files and using CSS3 media queries [2] to select the correct one
for the user display.

[1] http://wiki.sugarlabs.org/go/Human_Interface_Guidelines/The_Sugar_Interface/Layout_Guidelines
[2] http://www.w3.org/TR/css3-mediaqueries/