More on palettes
Only one palette visible per activity.  A palette popup provokes the
popdown of the current visible one, if any.

If the invoker is a toolbar button, add a background box on top of the
palette, over the button.

If the invoker is not a toolbar button, the palette will popup in the
clicked position.