Use more compatible way to create CustomEvent
As reported by Lionel:
"Like lot of HTML5 features, CustomEvent is only a draft.
BTW the version without constructor [1] is currently more deployed in browsers
than the one with a constructor [2].
The sugar-web template use the later in two files activity.js and menupalette.js.
It's possible to convert the call to the version with constructor to the version
without constructor without changing the way of working of the code."

This change allow the code run in Android devices.

[1] https://dvcs.w3.org/hg/dom3events/raw-file/tip/html/DOM3-Events.html#interface-CustomEvent
[2] https://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#interface-customevent

Email: http://lists.sugarlabs.org/archive/sugar-devel/2014-June/048420.html